requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.5
et_xmlfile==2.0.0
gunicorn==21.2.0
mysqlclient==2.1.1
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
watchdog==5.0.3
tenacity==8.5.0